Refactor code for improved readability and consistency
- Simplified error handling in vocab.json parsing in CtcGreedyDecoder.
- Enhanced formatting for debug messages in CtcGreedyDecoder and GigaAmModel.
- Cleaned up code structure in GigaAmMelExtractor and ParakeetModel for better clarity.
- Streamlined LstmLayer and PredictionNet implementations for consistency.
- Improved readability in JointNetwork and TdtGreedyDecoder by restructuring conditions.
- Updated WhisperModel logging for better clarity and consistency.
- Refactored Qwen3AsrModel for improved readability in quantization detection.
- General code cleanup across multiple files to enhance maintainability and readability.

+++ crates/aut-encoder/tests/load_encoder.rs
@@ -85,7 +85,7 @@ fn test_load_encoder_from_safetensors() {
                     // Для кратностей 100 это даёт 13 токенов на 100 mel-фреймов.
                     fn hf_aftercnn_len(input_len: usize) -> usize {
                         let input_len = input_len as i64;
-                        let leave = (input_len % 100) as i64;
+                        let leave = input_len % 100;
                         let feat = (leave - 1).div_euclid(2) + 1;
                         let out = (((feat - 1).div_euclid(2) + 1 - 1).div_euclid(2) + 1)
                             + (input_len / 100) * 13;
